requirements.txt
pillow>=6.2.0
pandas>=0.25.1
streamlit>=0.61.0
scikit-learn>=0.22.1